Create nlp_grading_essays.py
This is one of the files I created for my research assistant position. The purpose of this file is to create models that analyze grades of different categories for each essay and then analyze the accuracy of those models. More features will be added throughout the semester. The files used will not be posted, but they are formatted with the essays themselves - one per row - and grades for each category scored. Each essay was graded by one or two graders, so the match function was needed to completely fill in the imported data frame.