Create airplane_cpp.cpp
This was from my first computer science C++ programming class. It was my final project after three months of studying. I taught myself the course and tested out of the class. The purpose of the project was to create seating arrangements for an airplane and ask users to make selections for their seat.